DTOS, Services, Exceptions, Repos & Controllers creation.

+++ src/main/java/com/example/be_java_hisp_w25_g01/entity/Product.java
@@ -11,7 +11,7 @@ import lombok.experimental.FieldDefaults;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class Product {
-    long productId;
+    Integer productId;
     String productName;
     String type;
     String brand;
